[TCHL-41, TCHL-46] Improvements over the admin and user logout and minor enhancments

+++ ChatSQL/code/cliAdmin.py
@@ -31,8 +31,14 @@ def admin():
         elif re.match(r"^3$|^get$", choice):
             print(getFiles(database_path))
         elif re.match(r"^4$|^exit$", choice):
-            print("Leaving admin section. You will be redirected to the main menu.")
-            break
+            confirmation = input("Are you sure you want to exit the admin section? (yes/no): ").lower()
+            if confirmation == "yes":
+                print("Leaving admin section. You will be redirected to the main menu.")
+                break
+            elif confirmation == "no":
+                print("Returning to the admin section menu.")
+            else:
+                print("Invalid choice. Returning to the admin section menu.")
         else:
             print("Invalid choice")
 
@@ -83,11 +89,36 @@ def addFile():
 
 def deleteFile():
     filename_to_delete = input("Enter the filename you want to delete from the database: ")
-    file_path = os.path.join(database_path, filename_to_delete)
 
-    if os.path.exists(file_path):
-        os.remove(file_path)
-        loading_animation(.75)
-        print(f"File '{filename_to_delete}' has been deleted.")
+    # Aggiunta dell'estensione .json se l'utente non l'ha fornita
+    if not filename_to_delete.endswith(".json"):
+        filename_to_delete_with_extension = filename_to_delete + ".json"
+        filename_to_delete_without_extension = filename_to_delete
     else:
+        filename_to_delete_with_extension = filename_to_delete
+        filename_to_delete_without_extension = filename_to_delete.replace(".json", "")
+
+    # Tentativo di eliminare il file specificato con e senza estensione .json
+    file_paths_to_try = [
+        os.path.join(database_path, filename_to_delete_with_extension),
+        os.path.join(database_path, filename_to_delete_without_extension)
+    ]
+
+    file_deleted = False
+    for file_path in file_paths_to_try:
+        if os.path.exists(file_path):
+            # Chiedi conferma prima di eliminare il file
+            confirm = input(f"Are you sure you want to delete '{os.path.basename(file_path)}'? (yes/no): ").lower()
+            if confirm == 'yes':
+                # Elimina il file
+                os.remove(file_path)
+                print(f"File '{os.path.basename(file_path)}' has been deleted.")
+                file_deleted = True
+                break
+            else:
+                print(f"Deletion of '{os.path.basename(file_path)}' canceled.")
+                file_deleted = True  # Consider the file as deleted to skip the "not file_deleted" message
+                break
+
+    if not file_deleted:
         print(f"File '{filename_to_delete}' does not exist in the database directory.")
